[Archive] Pathfinder++ 0.06

+++ src/interface/pathinfo.h
@@ -2,14 +2,12 @@
 #define PATHINFO_H
 
 #include <SFML/Graphics.hpp>
-#include <sstream>
 
 class PathInfo {
 public:
 	PathInfo(sf::Font& font, sf::Vector2i pos);
 
-	void setInformations(int pathSize, float time, int cCount);
-
+	void setInformations(int pathSize, long long time, int cCount);
 	void draw(sf::RenderWindow& window);
 
 private:
